+++ src/main/java/com/brandon3055/draconicevolution/items/tools/DraconicStaffOfPower.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class DraconicStaffOfPower extends MiningToolBase {
     public DraconicStaffOfPower() {
         super(ToolStats.DRA_STAFF_ATTACK_DAMAGE, ToolStats.DRA_STAFF_ATTACK_SPEED, PICKAXE_OVERRIDES);//TODO Attack Damage and speed
-        this.baseMiningSpeed = ToolStats.DRA_STAFF_MINING_SPEED;
+        this.baseMiningSpeed = (float)ToolStats.DRA_STAFF_MINING_SPEED;
         this.baseAOE = ToolStats.BASE_DRACONIC_MINING_AOE + 1;
         setEnergyStats(ToolStats.DRACONIC_BASE_CAPACITY * 3, 16000000, 0);
 //        this.setHarvestLevel("all", 10);
